refactor(config): move document_helpers to dedicated file
